omolu formatting

+++ lib/Style/ExpressionsContext.hs
@@ -3,15 +3,16 @@
 module Style.ExpressionsContext where
 
 import Control.Lens
-import Proto.Vector_tile.Tile.Layer
 import Proto.Vector_tile.Tile.Feature
 import Proto.Vector_tile.Tile.GeomType
+import Proto.Vector_tile.Tile.Layer
 
 data ExpressionContext = ExpressionContext
-  { _feature :: Feature
-  , _layer   :: Layer
-  , _ctxZoom    :: Int
-  } deriving (Show, Eq)
+  { _feature :: Feature,
+    _layer :: Layer,
+    _ctxZoom :: Double
+  }
+  deriving (Show, Eq)
 
 makeLenses ''ExpressionContext
